simplify constructor.

+++ managementplatformv1 3/managementplatformv1/dataBackup6005/src/main/java/com/earthquake/managementPlatform/service/BackupService.java
@@ -19,12 +19,6 @@ import java.util.List;
 public class BackupService {
 
     @Resource
-    private RestTemplate restTemplate;
-
-    @Value("${storageInformation.url}")
-    private String storageInformationUrl;
-
-    @Resource
     BrickwoodStructureMapper brickwoodStructureMapper;
     @Resource
     CivilStructureMapper civilStructureMapper;
@@ -72,6 +66,10 @@ public class BackupService {
     BasicEarthquakeInfoMapper basicEarthquakeInfoMapper;
     @Resource
     DisasterPredictionMapper disasterPredictionMapper;
+    @Resource
+    private RestTemplate restTemplate;
+    @Value("${storageInformation.url}")
+    private String storageInformationUrl;
 
     @Transactional
     public boolean backUp(int time) {
@@ -79,7 +77,8 @@ public class BackupService {
 
             List<DeathStatistics> deathStatisticses = restTemplate.exchange(storageInformationUrl + "/v1/deathStatisticsCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<DeathStatistics>>() {
             }).getBody();
-            if (deathStatisticses.size() > 0) {
+
+            if (deathStatisticses != null && deathStatisticses.size() > 0) {
                 for (DeathStatistics deathStatistics : deathStatisticses) {
                     deathStatisticsMapper.save(deathStatistics);
                     restTemplate.delete(storageInformationUrl + "/v1/deathStatistics/" + deathStatistics.getId());
@@ -88,7 +87,7 @@ public class BackupService {
 
             List<InjuredStatistics> injuredStatisticses = restTemplate.exchange(storageInformationUrl + "/v1/injuredStatisticsCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<InjuredStatistics>>() {
             }).getBody();
-            if (injuredStatisticses.size() > 0) {
+            if (injuredStatisticses != null && injuredStatisticses.size() > 0) {
                 for (InjuredStatistics injuredStatistics : injuredStatisticses) {
                     injuredStatisticsMapper.save(injuredStatistics);
                     restTemplate.delete(storageInformationUrl + "/v1/injuredStatistics/" + injuredStatistics.getId());
@@ -97,7 +96,7 @@ public class BackupService {
 
             List<MissingStatistics> missingStatisticses = restTemplate.exchange(storageInformationUrl + "/v1/missingStatisticsCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<MissingStatistics>>() {
             }).getBody();
-            if (missingStatisticses.size() > 0) {
+            if (missingStatisticses != null && missingStatisticses.size() > 0) {
                 for (MissingStatistics missingStatistics : missingStatisticses) {
                     missingStatisticsMapper.save(missingStatistics);
                     restTemplate.delete(storageInformationUrl + "/v1/missingStatistics/" + missingStatistics.getId());
@@ -106,7 +105,7 @@ public class BackupService {
 
             List<CivilStructure> civilStructures = restTemplate.exchange(storageInformationUrl + "/v1/civilStructureCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<CivilStructure>>() {
             }).getBody();
-            if (civilStructures.size() > 0) {
+            if (civilStructures != null && civilStructures.size() > 0) {
                 for (CivilStructure civilStructure : civilStructures) {
                     civilStructureMapper.save(civilStructure);
                     restTemplate.delete(storageInformationUrl + "/v1/civilStructure/" + civilStructure.getId());
@@ -115,7 +114,7 @@ public class BackupService {
 
             List<MasonryStructure> masonryStructures = restTemplate.exchange(storageInformationUrl + "/v1/masonryStructureCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<MasonryStructure>>() {
             }).getBody();
-            if (masonryStructures.size() > 0) {
+            if (masonryStructures != null && masonryStructures.size() > 0) {
                 for (MasonryStructure masonryStructure : masonryStructures) {
                     masonryStructureMapper.save(masonryStructure);
                     restTemplate.delete(storageInformationUrl + "/v1/masonryStructure/" + masonryStructure.getId());
@@ -124,7 +123,7 @@ public class BackupService {
 
             List<BrickwoodStructure> brickwoodStructures = restTemplate.exchange(storageInformationUrl + "/v1/brickwoodStructureCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<BrickwoodStructure>>() {
             }).getBody();
-            if (brickwoodStructures.size() > 0) {
+            if (brickwoodStructures != null && brickwoodStructures.size() > 0) {
                 for (BrickwoodStructure brickwoodStructure : brickwoodStructures) {
                     brickwoodStructureMapper.save(brickwoodStructure);
                     restTemplate.delete(storageInformationUrl + "/v1/brickwoodStructure/" + brickwoodStructure.getId());
@@ -133,7 +132,7 @@ public class BackupService {
 
             List<FrameworkStructure> frameworkStructures = restTemplate.exchange(storageInformationUrl + "/v1/frameworkStructureCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<FrameworkStructure>>() {
             }).getBody();
-            if (frameworkStructures.size() > 0) {
+            if (frameworkStructures != null && frameworkStructures.size() > 0) {
                 for (FrameworkStructure frameworkStructure : frameworkStructures) {
                     frameworkStructureMapper.save(frameworkStructure);
                     restTemplate.delete(storageInformationUrl + "/v1/frameworkStructure/" + frameworkStructure.getId());
@@ -142,7 +141,7 @@ public class BackupService {
 
             List<OtherStructure> otherStructures = restTemplate.exchange(storageInformationUrl + "/v1/otherStructureCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<OtherStructure>>() {
             }).getBody();
-            if (otherStructures.size() > 0) {
+            if (otherStructures != null && otherStructures.size() > 0) {
                 for (OtherStructure otherStructure : otherStructures) {
                     otherStructureMapper.save(otherStructure);
                     restTemplate.delete(storageInformationUrl + "/v1/otherStructure/" + otherStructure.getId());
@@ -151,7 +150,7 @@ public class BackupService {
 
             List<TrafficDisaster> trafficDisasters = restTemplate.exchange(storageInformationUrl + "/v1/trafficDisasterCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<TrafficDisaster>>() {
             }).getBody();
-            if (trafficDisasters.size() > 0) {
+            if (trafficDisasters != null && trafficDisasters.size() > 0) {
                 for (TrafficDisaster trafficDisaster : trafficDisasters) {
                     trafficDisasterMapper.save(trafficDisaster);
                     restTemplate.delete(storageInformationUrl + "/v1/trafficDisaster/" + trafficDisaster.getId());
@@ -160,7 +159,7 @@ public class BackupService {
 
             List<WaterDisaster> waterDisasters = restTemplate.exchange(storageInformationUrl + "/v1/waterDisasterCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<WaterDisaster>>() {
             }).getBody();
-            if (waterDisasters.size() > 0) {
+            if (waterDisasters != null && waterDisasters.size() > 0) {
                 for (WaterDisaster waterDisaster : waterDisasters) {
                     waterDisasterMapper.save(waterDisaster);
                     restTemplate.delete(storageInformationUrl + "/v1/waterDisaster/" + waterDisaster.getId());
@@ -169,7 +168,7 @@ public class BackupService {
 
             List<OilDisaster> oilDisasters = restTemplate.exchange(storageInformationUrl + "/v1/oilDisasterCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<OilDisaster>>() {
             }).getBody();
-            if (oilDisasters.size() > 0) {
+            if (oilDisasters != null && oilDisasters.size() > 0) {
                 for (OilDisaster oilDisaster : oilDisasters) {
                     oilDisasterMapper.save(oilDisaster);
                     restTemplate.delete(storageInformationUrl + "/v1/oilDisaster/" + oilDisaster.getId());
@@ -178,7 +177,7 @@ public class BackupService {
 
             List<GasDisaster> gasDisasters = restTemplate.exchange(storageInformationUrl + "/v1/gasDisasterCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<GasDisaster>>() {
             }).getBody();
-            if (gasDisasters.size() > 0) {
+            if (gasDisasters != null && gasDisasters.size() > 0) {
                 for (GasDisaster gasDisaster : gasDisasters) {
                     gasDisasterMapper.save(gasDisaster);
                     restTemplate.delete(storageInformationUrl + "/v1/gasDisaster/" + gasDisaster.getId());
@@ -187,7 +186,7 @@ public class BackupService {
 
             List<PowerDisaster> powerDisasters = restTemplate.exchange(storageInformationUrl + "/v1/powerDisasterCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<PowerDisaster>>() {
             }).getBody();
-            if (powerDisasters.size() > 0) {
+            if (powerDisasters != null && powerDisasters.size() > 0) {
                 for (PowerDisaster powerDisaster : powerDisasters) {
                     powerDisasterMapper.save(powerDisaster);
                     restTemplate.delete(storageInformationUrl + "/v1/powerDisaster/" + powerDisaster.getId());
@@ -196,7 +195,7 @@ public class BackupService {
 
             List<CommDisaster> commDisasters = restTemplate.exchange(storageInformationUrl + "/v1/commDisasterCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<CommDisaster>>() {
             }).getBody();
-            if (commDisasters.size() > 0) {
+            if (commDisasters != null && commDisasters.size() > 0) {
                 for (CommDisaster commDisaster : commDisasters) {
                     commDisasterMapper.save(commDisaster);
                     restTemplate.delete(storageInformationUrl + "/v1/commDisaster/" + commDisaster.getId());
@@ -205,7 +204,7 @@ public class BackupService {
 
             List<IrrigationDisaster> irrigationDisasters = restTemplate.exchange(storageInformationUrl + "/v1/irrigationDisasterCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<IrrigationDisaster>>() {
             }).getBody();
-            if (irrigationDisasters.size() > 0) {
+            if (irrigationDisasters != null && irrigationDisasters.size() > 0) {
                 for (IrrigationDisaster irrigationDisaster : irrigationDisasters) {
                     irrigationDisasterMapper.save(irrigationDisaster);
                     restTemplate.delete(storageInformationUrl + "/v1/irrigationDisaster/" + irrigationDisaster.getId());
@@ -214,7 +213,7 @@ public class BackupService {
 
             List<CollapseRecord> collapseRecords = restTemplate.exchange(storageInformationUrl + "/v1/collapseRecordCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<CollapseRecord>>() {
             }).getBody();
-            if (collapseRecords.size() > 0) {
+            if (collapseRecords != null && collapseRecords.size() > 0) {
                 for (CollapseRecord collapseRecord : collapseRecords) {
                     collapseRecordMapper.save(collapseRecord);
                     restTemplate.delete(storageInformationUrl + "/v1/collapseRecord/" + collapseRecord.getId());
@@ -223,7 +222,7 @@ public class BackupService {
 
             List<LandslideRecord> landslideRecords = restTemplate.exchange(storageInformationUrl + "/v1/landslideRecordCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<LandslideRecord>>() {
             }).getBody();
-            if (landslideRecords.size() > 0) {
+            if (landslideRecords != null && landslideRecords.size() > 0) {
                 for (LandslideRecord landslideRecord : landslideRecords) {
                     landslideRecordMapper.save(landslideRecord);
                     restTemplate.delete(storageInformationUrl + "/v1/landslideRecord/" + landslideRecord.getId());
@@ -232,7 +231,7 @@ public class BackupService {
 
             List<DebrisRecord> debrisRecords = restTemplate.exchange(storageInformationUrl + "/v1/debrisRecordCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<DebrisRecord>>() {
             }).getBody();
-            if (debrisRecords.size() > 0) {
+            if (debrisRecords != null && debrisRecords.size() > 0) {
                 for (DebrisRecord debrisRecord : debrisRecords) {
                     debrisRecordMapper.save(debrisRecord);
                     restTemplate.delete(storageInformationUrl + "/v1/debrisRecord/" + debrisRecord.getId());
@@ -241,7 +240,7 @@ public class BackupService {
 
             List<KarstRecord> karstRecords = restTemplate.exchange(storageInformationUrl + "/v1/karstRecordCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<KarstRecord>>() {
             }).getBody();
-            if (karstRecords.size() > 0) {
+            if (karstRecords != null && karstRecords.size() > 0) {
                 for (KarstRecord karstRecord : karstRecords) {
                     karstRecordMapper.save(karstRecord);
                     restTemplate.delete(storageInformationUrl + "/v1/karstRecord/" + karstRecord.getId());
@@ -250,7 +249,7 @@ public class BackupService {
 
             List<CrackRecord> crackRecords = restTemplate.exchange(storageInformationUrl + "/v1/crackRecordCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<CrackRecord>>() {
             }).getBody();
-            if (crackRecords.size() > 0) {
+            if (crackRecords != null && crackRecords.size() > 0) {
                 for (CrackRecord crackRecord : crackRecords) {
                     crackRecordMapper.save(crackRecord);
                     restTemplate.delete(storageInformationUrl + "/v1/crackRecord/" + crackRecord.getId());
@@ -259,7 +258,7 @@ public class BackupService {
 
             List<SettlementRecord> settlementRecords = restTemplate.exchange(storageInformationUrl + "/v1/settlementRecordCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<SettlementRecord>>() {
             }).getBody();
-            if (settlementRecords.size() > 0) {
+            if (settlementRecords != null && settlementRecords.size() > 0) {
                 for (SettlementRecord settlementRecord : settlementRecords) {
                     settlementRecordMapper.save(settlementRecord);
                     restTemplate.delete(storageInformationUrl + "/v1/settlementRecord/" + settlementRecord.getId());
@@ -268,7 +267,7 @@ public class BackupService {
 
             List<OtherRecord> otherRecords = restTemplate.exchange(storageInformationUrl + "/v1/otherRecordCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<OtherRecord>>() {
             }).getBody();
-            if (otherRecords.size() > 0) {
+            if (otherRecords != null && otherRecords.size() > 0) {
                 for (OtherRecord otherRecord : otherRecords) {
                     otherRecordMapper.save(otherRecord);
                     restTemplate.delete(storageInformationUrl + "/v1/otherRecord/" + otherRecord.getId());
@@ -277,7 +276,7 @@ public class BackupService {
 
             List<DisasterPrediction> disasterPredictions = restTemplate.exchange(storageInformationUrl + "/v1/disasterPredictionCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<DisasterPrediction>>() {
             }).getBody();
-            if (disasterPredictions.size() > 0) {
+            if (disasterPredictions != null && disasterPredictions.size() > 0) {
                 for (DisasterPrediction disasterPrediction : disasterPredictions) {
                     disasterPredictionMapper.save(disasterPrediction);
                     restTemplate.delete(storageInformationUrl + "/v1/disasterPrediction/" + disasterPrediction.getD_ID() + '/' + disasterPrediction.getS_ID());
@@ -286,19 +285,15 @@ public class BackupService {
 
             List<BasicEarthquakeInfo> basicEarthquakeInfos = restTemplate.exchange(storageInformationUrl + "/v1/disasterInfoCopy/" + time, HttpMethod.GET, null, new ParameterizedTypeReference<List<BasicEarthquakeInfo>>() {
             }).getBody();
-            if (basicEarthquakeInfos.size() > 0) {
+            if (basicEarthquakeInfos != null && basicEarthquakeInfos.size() > 0) {
                 for (BasicEarthquakeInfo basicEarthquakeInfo : basicEarthquakeInfos) {
                     basicEarthquakeInfoMapper.save(basicEarthquakeInfo);
                     restTemplate.delete(storageInformationUrl + "/v1/disasterInfo/" + basicEarthquakeInfo.getId());
                 }
             }
-
             return true;
         } catch (Exception e) {
             return false;
         }
-
-
     }
-
 }
